refactor : dto
