+++ FashionToday_Back/src/main/java/hamburger/fashiontoday/controller/UserController.java
@@ -2,7 +2,7 @@ package hamburger.fashiontoday.controller;
 
 import hamburger.fashiontoday.domain.member.Member;
 import hamburger.fashiontoday.domain.member.MemberRepository;
-import hamburger.fashiontoday.dto.UserInfo;
+import hamburger.fashiontoday.domain.member.MemberInfo;
 import hamburger.fashiontoday.service.JwtService;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -41,23 +41,23 @@ public class UserController {
      * @return
      */
     @PostMapping("/userInfo")
-    public UserInfo userInfo(@RequestHeader(value = "Authorization")String authorization){
+    public MemberInfo userInfo(@RequestHeader(value = "Authorization")String authorization){
 
         System.out.println("나의 토큰 :"+authorization);
 
         if(jwtService.isUsable(authorization)){
             System.out.println("유저 아이디 : "+jwtService.getMember(authorization));
             Member member = memberRepository.findByMId(3);
-            UserInfo userInfo = new UserInfo(member.getMId(),member.getMName(),member.getMStar(),member.getMProfileUrl(),10,"200");
-            logger.debug(programId + " : userInfo - success : memberId = "+member.getMId());
-            return userInfo;
+            MemberInfo memberInfo = new MemberInfo(member.getMId(),member.getMName(),member.getMStar(),member.getMProfileUrl(),10,"200");
+            logger.debug(programId + " : memberInfo - success : memberId = "+member.getMId());
+            return memberInfo;
         }
 
         System.out.println("토큰 안됨");
         Member member = memberRepository.findByMId(3);
-        UserInfo userInfo = new UserInfo(member.getMId(),member.getMName(),member.getMStar(),member.getMProfileUrl(),10,"200");
-        logger.debug(programId + " : userInfo - success : memberId = "+member.getMId());
-        return userInfo;
+        MemberInfo memberInfo = new MemberInfo(member.getMId(),member.getMName(),member.getMStar(),member.getMProfileUrl(),10,"200");
+        logger.debug(programId + " : memberInfo - success : memberId = "+member.getMId());
+        return memberInfo;
     }
