+++ FashionToday_Back/src/main/java/hamburger/fashiontoday/domain/lookitem/LookitemInfo.java
@@ -1,0 +1,22 @@
+package hamburger.fashiontoday.domain.lookitem;
+
+
+import lombok.Getter;
+import lombok.Setter;
+
+/**
+ * @프로그램ID : HAM-PB-3003-J
+ * @프로그램명 : LookitemInfo.java
+ * @author : 심기성
+ * @date : 2019.10.15
+ * @version : 0.5
+ *
+ */
+
+@Getter
+@Setter
+public class LookitemInfo {
+
+
+
+}
