+++ FashionToday_Back/src/main/java/hamburger/fashiontoday/domain/member/MemberInfo.java
@@ -1,4 +1,4 @@
-package hamburger.fashiontoday.dto;
+package hamburger.fashiontoday.domain.member;
 
 import lombok.Getter;
 import lombok.Setter;
@@ -6,7 +6,7 @@ import lombok.Setter;
 
 /**
  * @프로그램ID : HAM-PB-3001-J
- * @프로그램명 : UserInfo.java
+ * @프로그램명 : MemberInfo.java
  * @author : 심기성
  * @date : 2019.09.01
  * @version : 0.5
@@ -14,7 +14,7 @@ import lombok.Setter;
  */
 @Getter
 @Setter
-public class UserInfo {
+public class MemberInfo {
 
     private int user_id;
 
@@ -29,7 +29,7 @@ public class UserInfo {
     private String apiseq;
 
 
-    public UserInfo(int user_id, String name, int star, String profile_image, int msec, String apiseq) {
+    public MemberInfo(int user_id, String name, int star, String profile_image, int msec, String apiseq) {
         this.user_id = user_id;
         this.name = name;
         this.star = star;
